deployment/requirements.txt
aiohttp>=3.9.0
aiohttp-cors==0.8.0
websockets==11.0.3
tensorflow==2.13.0
numpy==1.24.3
scikit-learn==1.3.0
gunicorn==21.2.0
python-dotenv==0.19.0 